// src/app/modelos/chofer.interface.ts
export interface Chofer {
	idChofer?: number;	
	cveMatriculaChofer?: string;	
	nombreChofer?: string;	
	cveUnidadAdscripcion?: string;	
	cveUnidadOOAD?: string;	
	fecInicioContrato?: string;	
	fecFinContrato?: string;	
	fecIniIncapacidad?: string;	
	idUnidadAdscripcion?: number;	
	fecFinIncapacidad?: string;	
	desCategoria?: string;	
	desMotivo?: string;	
	cveLicencia?: string;	
	cveTipoLicencia?: string;	
	fecVigencia?: string;	
	fecExpedicion?: string;	
	desrutaLicencia?: string;	
	estatusChofer?: string;	
	cveMatricula?: string;	
	fecAlta?: string;	
	fecActualizacion?: string;	
	fecBaja?: string;	
	indActivo?: number;	
	indSistema?: number;
}